requirements.txt
# Core LLM & Agent Framework
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langchain-chroma>=0.2.0

# Embeddings
sentence-transformers>=3.0.0

# Vector Store
chromadb>=0.5.0

# Document Processing
arxiv>=2.1.0
pypdf>=4.0.0

# Web Search
duckduckgo-search>=6.0.0

# API
fastapi>=0.115.0
uvicorn>=0.30.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.27.0

# Results & Visualization
matplotlib>=3.8.0
tabulate>=0.9.0
